+++ ListableUI/Sources/Internal/PresentationState/PresentationState.ItemState.swift
@@ -36,7 +36,12 @@ protocol AnyPresentationItemState : AnyObject
     
     func applyToVisibleCell(with environment : ListEnvironment)
         
-    func setNew(item anyItem : AnyItem, reason : PresentationState.ItemUpdateReason, updateCallbacks : UpdateCallbacks)
+    func set(
+        new : AnyItem,
+        reason : PresentationState.ItemUpdateReason,
+        updateCallbacks : UpdateCallbacks,
+        environment : ListEnvironment
+    )
     
     func willDisplay(cell : UICollectionViewCell, in collectionView : UICollectionView, for indexPath : IndexPath)
     func didEndDisplay()
@@ -170,10 +175,16 @@ extension PresentationState
                     return
                 }
                 
-                self.setNew(item: new, reason: .updateFromItemCoordinator, updateCallbacks: UpdateCallbacks(.immediate))
+                let environment = dependencies.environmentProvider()
+                
+                self.set(
+                    new: new,
+                    reason: .updateFromItemCoordinator,
+                    updateCallbacks: UpdateCallbacks(.immediate, wantsAnimations: true),
+                    environment: environment
+                )
                 
                 animation.perform {
-                    self.applyToVisibleCell(with: dependencies.environmentProvider())
                     delegate.coordinatorUpdated(for: self.anyModel)
                 }
             }
@@ -303,10 +314,15 @@ extension PresentationState
             )
         }
         
-        func setNew(item anyItem: AnyItem, reason: ItemUpdateReason, updateCallbacks : UpdateCallbacks)
+        func set(
+            new : AnyItem,
+            reason : PresentationState.ItemUpdateReason,
+            updateCallbacks : UpdateCallbacks,
+            environment : ListEnvironment
+        )
         {
             let old = self.model
-            let new = anyItem as! Item<Content>
+            let new = new as! Item<Content>
             
             self.storage.model = new
             
@@ -338,6 +354,17 @@ extension PresentationState
 
             if reason != .noChange {
                 self.resetCachedSizes()
+            }
+            
+            let wantsReapplication = self.model.reappliesToVisibleView.shouldReapply(
+                comparing: old.reappliesToVisibleView,
+                isEquivalent: reason == .noChange
+            )
+            
+            if wantsReapplication {
+                updateCallbacks.performAnimation {
+                    self.applyToVisibleCell(with: environment)
+                }
             }
         }
         
@@ -457,7 +484,7 @@ extension PresentationState
                     self.applyTo(
                         cell: cell,
                         itemState: itemState,
-                        reason: .willDisplay,
+                        reason: .measurement,
                         environment: environment
                     )